+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## v9.1.0 *(2023-09-?)*
+## v9.1.0 *(2023-09-26)*
 - New components: `InputRating`, `Tag`, `InputTag`, `InputTime`.
 - Add `hideTip` and more, to `Popover`.
 - Small bugfixes and improvements.